Added factory class
